Decouple tunnel parameters generation from daemon message handler
